+++ app/scripts/compiled-es5/storage.js
@@ -28,7 +28,13 @@ _module.service('cookieHandler', ['$cookies', function ($cookies) {
     };
 }]);
 
-_module.factory('languageStorage', ['$rootScope', '$window', 'languageSelectConfig', 'cookieHandler', function ($rootScope, $window, languageSelectConfig, cookieHandler) {
+_module.factory('windowReload', [function () {
+    return function () {
+        console.log('a');
+    };
+}]);
+
+_module.factory('languageStorage', ['$rootScope', '$window', 'languageSelectConfig', 'cookieHandler', 'windowReload', function ($rootScope, $window, languageSelectConfig, cookieHandler, windowReload) {
     var cookieSignature = 'selectedLanguage';
     var eventSignature = 'language-select:change';
 
@@ -85,7 +91,7 @@ _module.factory('languageStorage', ['$rootScope', '$window', 'languageSelectConf
         }
     };
 
-    var determineStartingLanguageId = function determineStartingLanguageId() {
+    var determineStartingLanguage = function determineStartingLanguage() {
         var rawCookieLanguageId = cookieHandler.get(cookieSignature);
         var rawBrowserLanguageId = $window.navigator.language || $window.navigator.userLanguage;
 
@@ -93,11 +99,18 @@ _module.factory('languageStorage', ['$rootScope', '$window', 'languageSelectConf
         var browserLanguageId = getLanguageIdIfValid(rawBrowserLanguageId);
         var defaultLanguageId = languageSelectConfig.defaultLanguageId();
 
-        return cookieLangaugeId || browserLanguageId || defaultLanguageId;
+        return {
+            id: cookieLangaugeId || browserLanguageId || defaultLanguageId,
+            cookieWasSet: Boolean(rawCookieLanguageId)
+        };
     };
 
-    var startingLanguageId = determineStartingLanguageId();
-    publicMethods.set(startingLanguageId);
+    var startingLanguage = determineStartingLanguage();
+    publicMethods.set(startingLanguage.id);
+
+    if (!startingLanguage.cookieWasSet) {
+        windowReload();
+    }
 
     return publicMethods;
 }]);